src/main/java/net/amygdalum/comtemplate/engine/resolvers/EqualsResolver.java
package net.amygdalum.comtemplate.engine.resolvers;

import static java.util.Arrays.asList;
import static net.amygdalum.comtemplate.engine.expressions.BooleanLiteral.bool;

import java.util.List;

import net.amygdalum.comtemplate.engine.Scope;
import net.amygdalum.comtemplate.engine.TemplateImmediateExpression;

public class EqualsResolver extends FunctionResolver {

	public EqualsResolver() {
		super("equals", 1);
	}

	@Override
	public TemplateImmediateExpression resolve(TemplateImmediateExpression base, List<TemplateImmediateExpression> arguments, Scope scope) {
		String actual = base.getText();
		String expected = arguments.get(0).getText();
		return bool(actual.equals(expected));
	}

	@Override
	public List<Class<? extends TemplateImmediateExpression>> getResolvedClasses() {
		return asList(TemplateImmediateExpression.class);
	}
}
